Update Spring Boot dependencies and refactor session controller
