+++ src/main/java/com/code/config/WebMvcJspApplication.java
@@ -23,14 +23,14 @@ import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.jdbc.DataSourceBuilder;
+import org.springframework.boot.jdbc.DataSourceBuilder;
 import org.springframework.boot.builder.SpringApplicationBuilder;
-import org.springframework.boot.context.embedded.tomcat.TomcatConnectorCustomizer;
-import org.springframework.boot.context.embedded.tomcat.TomcatEmbeddedServletContainerFactory;
+import org.springframework.boot.web.embedded.tomcat.TomcatConnectorCustomizer;
+import org.springframework.boot.web.embedded.tomcat.TomcatServletWebServerFactory;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.boot.context.web.SpringBootServletInitializer;
-import org.springframework.boot.orm.jpa.EntityScan;
+import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
+import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
@@ -89,9 +89,9 @@ public class WebMvcJspApplication extends SpringBootServletInitializer {
 	 //Tomcat large file upload connection reset
 	    //http://www.mkyong.com/spring/spring-file-upload-and-connection-reset-issue/
 	    @Bean
-	    public TomcatEmbeddedServletContainerFactory tomcatEmbedded() {
+	    public TomcatServletWebServerFactory tomcatEmbedded() {
 
-	        TomcatEmbeddedServletContainerFactory tomcat = new TomcatEmbeddedServletContainerFactory();
+	        TomcatServletWebServerFactory tomcat = new TomcatServletWebServerFactory();
 
 	        tomcat.addConnectorCustomizers((TomcatConnectorCustomizer) connector -> {
 	            if ((connector.getProtocolHandler() instanceof AbstractHttp11Protocol<?>)) {
